Add 'author is has any book' control for removeCmd

+++ WebApi/Application/AuthorOperations/Commands/DeleteAuthor/DeleteAuthorCommand.cs
@@ -2,6 +2,7 @@ using System;
 using System.Linq;
 using AutoMapper;
 using WebApi.DBOperations;
+using WebApi.Entities;
 
 namespace WebApi.Application.AuthorOperations.Commands.DeleteAuthor
 {
@@ -20,6 +21,11 @@ namespace WebApi.Application.AuthorOperations.Commands.DeleteAuthor
 
             if (author is null)
                 throw new InvalidOperationException("Yazar bulunmadı.");
+
+            var isHasAnyBook = _context.Books.FirstOrDefault(x => x.AuthorId == author.Id);
+
+            if (isHasAnyBook is not null)
+                throw new InvalidOperationException("Yazara ait kitap varken işlem gerçekleşemez.");
 
             _context.Authors.Remove(author);
             _context.SaveChanges();
